--- a/benchmark/results.txt
+++ b/benchmark/results.txt
@@ -1,4 +1,4 @@
-Benchmark run on Sat 12 Jan 2019 13:33:09 GMT with Node v10.13.0
+Benchmark run on Wed 16 Jan 2019 17:47:28 GMT with Node v10.13.0
 
 Using:
  - hyperapp@1.2.9
@@ -11,12 +11,13 @@ Using:
  - vdo@4.2.0
 
 Results:
- - Hyperapp x 7,729 ops/sec ±1.41% (89 runs sampled)
- - Hyperons x 9,393 ops/sec ±0.60% (95 runs sampled)
- - Inferno x 9,558 ops/sec ±0.31% (95 runs sampled)
- - Nerv x 4,880 ops/sec ±0.84% (94 runs sampled)
- - Preact x 4,418 ops/sec ±0.15% (96 runs sampled)
- - Rax x 9,731 ops/sec ±0.59% (93 runs sampled)
- - React x 5,020 ops/sec ±1.12% (93 runs sampled)
- - vdo x 6,234 ops/sec ±0.22% (96 runs sampled)
-The fastest is Rax
+ - Hyperapp x 7,577 ops/sec ±1.72% (92 runs sampled)
+ - Hyperons x 11,917 ops/sec ±1.89% (92 runs sampled)
+ - Inferno x 9,375 ops/sec ±0.67% (91 runs sampled)
+ - Nerv x 4,872 ops/sec ±0.73% (93 runs sampled)
+ - Preact x 4,332 ops/sec ±1.12% (96 runs sampled)
+ - Rax x 9,519 ops/sec ±1.15% (94 runs sampled)
+ - React x 5,020 ops/sec ±1.14% (92 runs sampled)
+ - vdo x 6,135 ops/sec ±0.62% (93 runs sampled)
+
+The fastest is: [ 'Hyperons' ]