--- a/src/create-element.js
+++ b/src/create-element.js
@@ -1,7 +1,8 @@
 function createElement(type, props, ...children) {
   props = props || {}
 
-  props.children = children || []
+  props.children =
+    Array.isArray(props.children) && children.length === 0 ? props.children : children
 
   // TODO: support default props
   return { type, props }