--- a/src/fragment.js
+++ b/src/fragment.js
@@ -1,0 +1,2 @@
+const Fragment = Symbol('Fragment')
+export default Fragment